feat: ProductUpdateVC 데이터 모델 프로토콜 추상화 #7
